Web modals now scroll to top upon navigation
