+++ homepairs/homepairsUI/src/modals/property-modals/EditApplianceModal/EditApplianceModal.tsx
@@ -221,7 +221,7 @@ export class EditApplianceModalBase extends React.Component<Props, EditState> {
     goBackToPreviousPage() {
         const { navigation } = this.props;
         const propId = navigation.getParam('propId');
-        navigation.replace(SingleProperty, { propId });
+        navigation.resolveModalReplaceNavigation(SingleProperty, { propId });
     }
 
     setInitialState() {
